+++ psims/test/test_transform_mzml.py
@@ -2,10 +2,10 @@ from io import BytesIO
 from psims.transform import mzml, utils
 
 from psims.test.test_data import datafile
-
+from psims.test.utils import UnclosableBuffer
 
 def test_mzml_pipe():
-    buff = BytesIO()
+    buff = UnclosableBuffer()
     path = datafile("small.mzML")
     st = mzml.MzMLTransformer(path, buff)
     st.write()
@@ -22,7 +22,7 @@ def test_mzml_pipe():
     for ref_spec, test_spec in zip(ref_reader.iterfind("fileDescription"),
                                    test_reader.iterfind("fileDescription")):
         assert utils.differ(ref_spec, test_spec)
-
+    super(UnclosableBuffer, buff).close()
 
 if __name__ == '__main__':
     test_mzml_pipe()